atmosphere_unit_service/requirements.txt
protobuf==4.25.2
grpcio>=1.60.1
pika==1.3.2
pyyaml==6.0.0
grpcio-tools>=1.60.1
